pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.15",
    "google-genai>=1.38.0",
    "google-generativeai>=0.8.0",
    "openai>=1.107.3",
    "pandas>=2.3.2",
    "plotly>=6.3.0",
    "streamlit>=1.49.1",
    "spacy>=3.7.0",
]
